Role added
Role.java contains an enum that describes the various roles a user can
have within a channel. An id value corresponding to the enumerated
ordinal value is associated with the role, as well as a roleName. The
latter is uppercase to follow convention and can be retrieved via a
toString() method.