Modified Java Docs and changed members to HashMap
The Java Docs have been improved to include proper descriptions of
return types and parameters. Unnecessary annotations have been removed.

The membersList has been changed to a hash map that is comprised of a
user, which serves as the key and their role. If the size of the hash
map is 0, the user is assigned the role of OWNER. Otherwise, newly added
users are assigned roles of USER.