Modifications to Channel and its repo
The Channel class has been modified to fit new designs for the POJOs.
It's members list was changed to a hashmap and then back to a list.
However, @MapKeyEnumerated does not generate a column for the Users
as I had originially hoped. I'm going to look into this more.

The ChannelRepository has been modified accordingly.