DTO, Security, and Utils packages added
These packages serve to separate out responsiblity. Data Transfer
Objects (DTOs) contains classes that will be used to transport data.
Presently, it contains Credentials.java, which stores a User's username
and password.

The Security package contains a variety of classes related to security.
The only class is FieldValidator, and it contains methods to validate
user input. Later, this package will likely contain JWT classes as well
as a security annotation.

Utils contains variables that will be used throughout the application
and are annoying to constantly redefine in different files. RegexUtil
stores final Strings of the regular expressions for an RFC 5322
compliant email address and a valid username.