User Controller and modifictions [Unstable]
The User Controller has been populated with methods. Testing is still in
the works. I've commented out Account related lines in User.java in
preparation for removing Account from User and creating a User in
Account.java on the Dev branch.

UserRepos and UserServices have been refactored to return Users for add
and update methods instead of booleans and only status codes,
respectively.