Modifications to account for Role enum
The principal and authfilter formerly treated role as a string. This is
not ideal for enums because their internal constructors are private.
Upon receipt of an Role, Role.valueOf(param) is called to convert the
string received from the Angular side into a proper enum value.
Principal had Role converted to an enum value in its constructors.