General Registration Page Structure
The input boxes and formGroups have been added. There are errors now
because the component does not include validators or ngModules yet.

I'm not sure how to extend the height of the bio input field while
keeping the start of the text when typing aligned with the upper-left
corner. I might also look into country and state drop-down menus
instead of text entries.